CC-14662 Refactoring after CR
